_val_info has a performance bug with torch tensors
Inordinate amounts of time are spent in formatting torch tesnors
to strings. This is particularly problematic in the inception
example with large numbers of high dimensional tensors. This cannot
be alleviated by passing verbose=False to the EvaluableGraph because
string formatting of the tensors happens regardless of whether
they are printed. For now, I simple guarded all print statements
with if verbose, for now. We should probably rewrite this code
eventually to use logging and to not format tensors unless messages
are set to display. Not sure the cleanest way to do this.